added dao and some forms

+++ src/main/java/ru/rsreu/nis/database/dao/MarkDAO.java
@@ -1,13 +1,12 @@
 package ru.rsreu.nis.database.dao;
 
-import ru.rsreu.nis.data.Mark;
-import ru.rsreu.nis.data.Trip;
+import ru.rsreu.nis.entity.Mark;
 
 import java.util.List;
 
 public interface MarkDAO {
     List<Mark> findAll();
-    Mark getMarkById(int id);
+    Mark getMarkById(Long id);
     Mark getMarkById(String id);
     void updateMark(Mark mark);
     void deleteMark(Mark mark);
