+++ src/main/java/ru/rsreu/nis/database/dao/RoleDAO.java
@@ -1,11 +1,11 @@
 package ru.rsreu.nis.database.dao;
 
-import ru.rsreu.nis.data.Role;
+import ru.rsreu.nis.entity.Role;
 
 import java.util.List;
 
 public interface RoleDAO {
         List<Role> findAll();
-        Role getRoleById(int id);
+        Role getRoleById(Long id);
         Role getRoleById(String id);
 }
