+++ src/main/java/ru/rsreu/nis/database/dao/TripDAO.java
@@ -1,12 +1,12 @@
 package ru.rsreu.nis.database.dao;
 
-import ru.rsreu.nis.data.Trip;
+import ru.rsreu.nis.entity.Trip;
 
 import java.util.List;
 
 public interface TripDAO {
     List<Trip> findAll();
-    Trip getTripById(int id);
+    Trip getTripById(Long id);
     Trip getTripById(String id);
     void updateDevice(Trip trip);
     void deleteDevice(Trip trip);
